+++ R/humanize.R
@@ -10,7 +10,6 @@
 #' humanize("natator-depressus")
 #' humanize("successful-crawl")
 humanize <- function(urlsafe_string) {
-  . <- ""
   urlsafe_string %>%
     stringr::str_to_title(.) %>%
     stringr::str_replace_all("-", " ")